Add api.js file and exported at index.js file
