+++ src/index.html
@@ -7,7 +7,7 @@
     <title>Document</title>
   </head>
   <body>
-    <section class="hotels" id="hotels">
+    <section class="charct" id="charct">
       <div class="container">
         <ul class="charc-list" id="charc-list"></ul>
       </div>
