+++ src/styles.css
@@ -2,6 +2,10 @@ body {
   background-color: beige;
 }
 
+.container {
+  padding: 0 50px;
+}
+
 .charc-list{
   display: flex;
   gap: 20px;
@@ -11,8 +15,19 @@ body {
   padding: 0;
 }
 
-
 .charc-list > li {
 padding: 0;
 width: 216px;
+}
+
+.like {
+  background-color: transparent;
+  color: blue;
+  border: solid 1px blue;
+}
+
+.like:active {
+  background-color: darkcyan;
+  color: blue;
+  border: solid 1px blue;
 }
